requirements.txt
torch>=2.3.0
torchvision>=0.18.0
matplotlib>=3.8.0
numpy>=1.26.0
Pillow>=10.0.0
clean-fid>=0.1.35
kaggle>=1.6.17
